remove duplication in theme test helper functions
from 16 to 4 helper definitions
